engine: extract esm/layers/great_circle_route_layer.js
